requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
multitasking==0.0.11
nodeenv==1.8.0
numpy==1.26.4
pandas==2.2.1
peewee==3.17.1
prisma==0.12.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
resend==0.7.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
StrEnum==0.4.15
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.0
webencodings==0.5.1
yfinance==0.2.37
